fix: keep Job label out of the Worker jobs module
