+++ src/session-title.test.ts
@@ -1,7 +1,7 @@
 import assert from "node:assert/strict";
 import test from "node:test";
 import { deriveSessionTitle, isScheduledJobRunMessage } from "./session-title";
-import { SCHEDULED_JOB_RUN_PREFIX } from "./jobs";
+import { SCHEDULED_JOB_RUN_PREFIX } from "./jobs-prompt";
 
 test("scheduled job injects are not owner messages", () => {
   assert.equal(isScheduledJobRunMessage(`${SCHEDULED_JOB_RUN_PREFIX}\n\nCheck the status.`), true);
